% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hapsampler.R
\name{hapsampler}
\alias{hapsampler}
\title{Haplotype sampler}
\usage{
hapsampler(data = NULL, trait = NULL, probthresh = 0.95)
}
\arguments{
\item{data}{a data frame object generated by running ‘read_data’.}

\item{trait}{the column name in the data file of the trait. This must be
specified.}

\item{probthresh}{a probability value between 0 and 1. Haplotypes with
haplotype probabilities less than ‘probthresh’ are discarded.}
}
\value{
to be written
}
\description{
Perform haplotpe analysis
}
\details{
to be written
}
\examples{
complete.name <- system.file("extdata", "dataexample.dat", package="HapSampler")
    # read in phenotypic data which is in csv format
    dt <- read_data(path=dirname(complete.name),
                                 file=basename(complete.name))

    hapres <- hapsampler(data=dt, trait="NAVEL")
    print(hapres)
}

