% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hapsampler.R
\name{is.HS}
\alias{is.HS}
\title{HS}
\usage{
is.HS(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
A logical value is returned.
}
\description{
tests if object is a HapSampler list structure
}

