% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotEffects}
\alias{plotEffects}
\alias{ploteffects}
\title{Plot Effects}
\usage{
plotEffects(x)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}
}
\value{
a plotting window is opened.
}
\description{
A trace plot of the mean qtl effect sizes of QQ, Qq, and qq from 
      the running of \code{\link{hapsampler}}.
}
\details{
Produces a plot of the change in mean qtl effect size  for each chain, over 
the run length.  It is useful for looking for agreeement between chains and also 
that the mean effect size estimates are behaving.
}
\seealso{
\code{\link{plotLike}}
}

