% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotLike.HS}
\alias{plotLike.HS}
\alias{plotlike.HS}
\title{Plot LogLikelihood}
\usage{
\method{plotLike}{HS}(x)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}
}
\value{
a plotting window is opened.
}
\description{
A log likelihood trace plot of hapsampler output.
}
\details{
Produces a plot of the change in log likelihood for each chain, over 
the run length.  It is useful for looking for agreeement between chains and also 
that the log likelihoods have stabilized over a run.
}
\seealso{
\code{\link{plotEffects}}
}

