% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotLike}
\alias{plotLike}
\alias{plotlike}
\title{Plot LogLikelihood}
\usage{
plotLike(x)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}

\item{...}{additional plot options. (not yet implemented).}
}
\value{
a plotting window is opened.
}
\description{
A log likelihood trace plot of hapsampler output.
}
\details{
Produces a plot of the change in log likelihood for each chain, over
the run length.  It is useful for looking for agreeement between chains and also
that the log likelihoods have stabilized over a run.
}
\seealso{
\code{\link{plotEffects}}
}

