% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotscatter.HS}
\alias{plotscatter.HS}
\title{Haplotype and QTL Allele Scatter Plots}
\usage{
\method{plotscatter}{HS}(x, nbins)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}

\item{nbins}{the number of bins for dividing the sampled haplotype frequency.}
}
\value{
a plotting window is opened.
}
\description{
Various scatter plots involving the sampled haplotypes and sampled QTL alleles.
}
\details{
A single graphic window is open, containing three plots.

The top left  plot is of the sampled haplotypes within a chain verse the probability of that
haplotype carrying QTL allele Q.  The sammpled haplotypes are ordered based on their average
allele Q probability (Prob(Q)).  The results from each chain are coloured
differently.  This plot is useful for identifying those haplotypes that have been sampled differently
across the parallel chains  (i.e. we don't want points a long way away from the red fitted line).


The top right plot is of the sampled haplotypes across chains verse the probability of that
haplotype carrying QTL allele Q. The sampled haplotypes are colour coded, depending upon
frequency of being sampled across all n chains. The frequency ranges associated with the colours
are given in the legend.

The bottom left plot is a scatter plot of the sampled QTL allele Q verse
its probability.
}
\seealso{
\code{\link{plotLike}}
}

