% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotscatter}
\alias{plotscatter}
\title{Haplotype and Allele Scatter Plots}
\usage{
plotscatter(x)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}
}
\value{
a plotting window is opened.
}
\description{
Various scatter plots involving haplotypes and observed alleles (ordered by mean Prob(A))
verse Prob(A).
}
\details{
This function opens a graphic window that contains three plots. The first
plot is of the sampled haplotypes (ordered by Prob(A)) against Prob(A).  The second plot
is similar, except the scater plot is of the unique haplotypes, binned according to the levels
given in the legend. The third plot is of the QTL alleles, ordered by mean P....
Produces a plot of the change in mean effect size  for each chain, over
the run length.  It is useful for looking for agreeement between chains and also
that the mean effect size estimates are behaving.
}
\seealso{
\code{\link{plotLike}}
}

