% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hapsampler.R
\name{hapsampler}
\alias{hapsampler}
\title{Haplotype Sampler}
\usage{
hapsampler(data = NULL, trait = NULL, nchains = 3, runlength = 30,
  probthresh = 0.95)
}
\arguments{
\item{data}{a data frame object generated by running \code{\link{read_data}}.}

\item{trait}{the column name in the data file of the trait. This must be
specified.}

\item{nchains}{the number of parallel chains to be run.}

\item{runlength}{the run length of each chain}

\item{probthresh}{a probability value between 0 and 1. Haplotypes with
         haplotype probabilities less than \code{probthresh}  are discarded.}
}
\value{
A list like object of class HS is returned with elements
\itemize{
\item{"trait"}{trait name}
\item{"nchains"}{number of chains}
\item{"canonical.hap"}{used by plotting functions}
\item{"haplotypes"}{a data.frame object containing the sampled haplotypes,
the number of times the haplotype is sampled across the chains, the mean probability of
the haplotype including QTL allele Q as approximated by the MCMC run, and the probability of
the haplotype including  QTL allele Q for each chain. }
}
}
\description{
Marker-QTL haplotypes are sampled, and their probabilities calculated via
 MCMC methods.
}
\details{
This function implements a MCMC algorithm for sampling the space of marker-QTL genotypes,
conditional on the observed marker haplotypes for each animal, the trait values, and
a penetrance function (i.e. the probability of the trait given the putative QTL alleles).

To guard against convergence issues, parallel chains are run by setting
the argument \code{nchains} to some integer value greater than 1.
\code{runlength} is the run length of a chain.

There are several plotting functions (such as \code{\link{plotLike}},
\code{\link{plotEffects}}, and \code{\link{plotScatter}}) available for
viewing the results.
}
\examples{
complete.name <- system.file("extdata", "dataexample.dat", package="HapSampler")
    # read in phenotypic data which is space separated
    dt <- read_data(path=dirname(complete.name),
                                 file=basename(complete.name))

    #  perform analysis for the NAVEL trait
    hapres <- hapsampler(data=dt, trait="NAVEL", nchains=2, runlength=5)

    # a plot of the log likelihood for each chain against run length
    plotLike(hapres)
}

