% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapsampler.R
\name{hapsampler}
\alias{hapsampler}
\title{Haplotype Sampler}
\usage{
hapsampler(data = NULL, trait = NULL, nchains = 3, runlength = 30,
  probthresh = 0.95)
}
\arguments{
\item{data}{a data frame object generated by running \code{\link{read_data}}.}

\item{trait}{the column name in the data file of the trait. This must be
specified.}

\item{nchains}{the number of parallel chains to be run.}

\item{runlength}{the run length of each chain}

\item{probthresh}{a probability theshold value.  Data on animals whose haplotype probability is less
than \code{probthresh} will be ignored.}
}
\value{
A list like object of class HS is returned with elements
\itemize{
\item trait:  the trait name
\item nchains: the number of chains
\item canonical.hap: this is the haplotype that is the most common in the input data. 
\item haplotypes: a data.frame object containing
\itemize{
   \item Haplotype:  the haplotype index
   \item Count: the frequency of the haplotype in the input data file after data on those animals with a haplotype probability less than 
   \code{probthresh} have been removed.
   \item MeanProb(Q):  the probability of this haplotype carrying the Q qtl allele, averaged over the mean of the Q allele probabilities from each chain
    \item Prob(Q):Chain1:  the mean probability of this haplotype carrying the Q qtl allele, for chain 1. This field is repeated for each chain that is run.
}
}
}
\description{
Marker-QTL haplotypes are sampled, and their probabilities calculated via
 MCMC methods.
}
\details{
This function implements a MCMC algorithm for sampling the space of marker-QTL genotypes, 
conditional on the observed marker haplotypes for each animal, the trait values, and 
a penetrance function (i.e. the probability of the trait given the putative QTL alleles). 

To guard against convergence issues, parallel chains are run by setting 
the argument \code{nchains} to some integer value greater than 1. 
\code{runlength} is the run length of a chain. 

There are several plotting functions (such as \code{\link{plotLike}}, 
\code{\link{plotEffects}}, and \code{\link{plotScatter}}) available for 
viewing the results.
}
\examples{

    complete.name <- system.file("extdata", "dataexample.dat", package="HapSampler")
    # read in phenotypic data which is space separated
    dt <- read_data(path=dirname(complete.name),
                                 file=basename(complete.name))
    
    #  perform analysis for the NAVEL trait
    hapres <- hapsampler(data=dt, trait="NAVEL", nchains=2, runlength=5)
    
    # a plot of the log likelihood for each chain against run length
    plotLike(hapres)

}

