% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarise.R
\name{plotEffects.HS}
\alias{plotEffects.HS}
\alias{ploteffects.HS}
\title{Plot Effects}
\usage{
\method{plotEffects}{HS}(x)
}
\arguments{
\item{x}{a \code{HS} object, obtained from running \code{\link{hapsampler}}.}
}
\value{
a plotting window is opened.
}
\description{
A trace plot of the mean effect sizes of AA, AB, and BB from
      the running of \code{\link{hapsampler}}.
}
\details{
Produces a plot of the change in mean effect size  for each chain, over
the run length.  It is useful for looking for agreeement between chains and also
that the mean effect size estimates are behaving.
}
\seealso{
\code{\link{plotLike}}
}

