% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(path = getwd(), file = NULL, csv = FALSE)
}
\arguments{
\item{path}{a character vector containing the absolute path for where the
         phenotype file is located.  The default is that the data file
         is contained in the directory from which R is being run.}

\item{file}{a character vector containing the name of the data file.}

\item{csv}{a logical value. When \code{TRUE}, a csv file format is assumed.
When \code{FALSE}, a space separated format is assumed.  The
default is for the file to be space separaed.}
}
\value{
a data frame is returned of the data. The names of the columns
    will be as specified by the first row of the data file.
}
\description{
Read in data file.
}
\details{
Here, \code{\link{read_data}} reads in the ASCII data into the package. A
    space separated ASCII file with column headings is assumed. A csv
    file can also be read if \code{csv} is set to \code{TRUE}. Each row of the
    data file is the data for a seperate animal.  The data file is
    allowed to have any number of columns.

    Three columns are mandatory in order for ‘hapsampler’ to function
    correctly; the two haplotype indexes and a haplotype probability.
    These three columns must be named hap1, hap2, and prob where prob
    is the haplotype probability of the first haplotype (hap1). The
    function will stop if the column names hap1, hap2, and prob are
    not found.

    The columns hap1 and hap2 contain the haplotype indexes. These
    indexes can be any integer number.

    NOTE WHAT about missing data???
}
\examples{
# Read in  example phenotypic data from ./inst/extdata/

    # finding the full location of the phenotypic data
    complete.name <- system.file("extdata", "dataexample.dat", package="HapSampler")

    # read in phenotypic data which is in csv format
    dt <- read_data(path=dirname(complete.name),
                                 file=basename(complete.name))

     ## print a couple of lines of the data file
     head(dt)
}

